configipy/requirements.txt
PyYAML==4.1

